refactor(user-management): relocate sso.py to autobot_shared, cosmetic diff only (#12647)
