+++ autobot-slm-backend/user_management/models/sso.py
@@ -2,244 +2,20 @@
 # SPDX-License-Identifier: Apache-2.0
 # AutoBot - AI-Powered Automation Platform
 # Author: mrveiss
+"""Re-export shim — the implementation now lives in autobot_shared (#12647).
+
+This file's only diff from the backend's copy was cosmetic/documentation (an
+unused ``Optional`` import and ``Mapped[Optional[...]]`` style) — see
+``autobot_shared/user_management/models/sso.py`` for the reconciliation
+note. Kept as a shim, not deleted, so every existing
+``from user_management.models.sso import ...`` importer keeps working
+unchanged.
 """
-SSO Provider and User SSO Link Models
 
-Supports multiple SSO providers:
-- Enterprise: LDAP/AD, SAML 2.0, Microsoft Entra ID, Google Workspace
-- Social: Google OAuth, Facebook, GitHub
-"""
+from autobot_shared.user_management.models.sso import (  # noqa: F401
+    SSOProvider,
+    SSOProviderType,
+    UserSSOLink,
+)
 
-import uuid
-from datetime import datetime
-from enum import Enum
-from typing import TYPE_CHECKING
-
-from sqlalchemy import Boolean, DateTime, ForeignKey, String, UniqueConstraint
-from sqlalchemy.dialects.postgresql import JSONB, UUID
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-
-from autobot_shared.ssot_constants import CategoryDefaults
-from autobot_shared.time_utils import now_utc
-from user_management.models.base import Base
-
-if TYPE_CHECKING:
-    from user_management.models.organization import Organization
-    from user_management.models.user import User
-
-
-class SSOProviderType(str, Enum):
-    """SSO Provider types."""
-
-    # Enterprise SSO
-    LDAP = "ldap"
-    ACTIVE_DIRECTORY = "active_directory"
-    OKTA = "okta"
-    SAML = "saml"
-    MICROSOFT_ENTRA = "microsoft_entra"
-    GOOGLE_WORKSPACE = "google_workspace"
-
-    # Social Login (provider mode)
-    GOOGLE = "google"
-    FACEBOOK = "facebook"
-    GITHUB = "github"
-
-
-class SSOProvider(Base):
-    """
-    SSO Provider configuration.
-
-    Stores configuration for SSO providers. Sensitive credentials (client_secret, bind_password)
-    are encrypted in the SystemSecret table; config JSONB stores non-sensitive data and references.
-    Can be organization-specific or global (for social login in provider mode).
-    """
-
-    __tablename__ = "sso_providers"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid4,
-    )
-
-    # Organization (nullable for global/social providers)
-    org_id: Mapped[uuid.UUID | None] = mapped_column(
-        UUID(as_uuid=True),
-        ForeignKey("organizations.id", ondelete="CASCADE"),
-        nullable=True,
-        index=True,
-    )
-
-    # Provider type
-    provider_type: Mapped[str] = mapped_column(
-        String(50),
-        nullable=False,
-        index=True,
-    )
-
-    # Display name
-    name: Mapped[str] = mapped_column(
-        String(255),
-        nullable=False,
-    )
-
-    # Provider configuration (JSONB)
-    # Sensitive fields (client_secret, bind_password) are stored encrypted in SystemSecret table.
-    # This config contains non-sensitive data (client_id, endpoints, etc.) and secret references.
-    config: Mapped[dict] = mapped_column(
-        JSONB,
-        nullable=False,
-    )
-
-    # Status
-    is_active: Mapped[bool] = mapped_column(
-        Boolean,
-        default=True,
-        nullable=False,
-    )
-
-    # Whether this is a social login provider (vs enterprise SSO)
-    is_social: Mapped[bool] = mapped_column(
-        Boolean,
-        default=False,
-        nullable=False,
-    )
-
-    # Allow user creation via this provider
-    allow_user_creation: Mapped[bool] = mapped_column(
-        Boolean,
-        default=True,
-        nullable=False,
-    )
-
-    # Default role for users created via this provider
-    default_role: Mapped[str | None] = mapped_column(
-        String(100),
-        nullable=True,
-        default=CategoryDefaults.ROLE_USER,
-    )
-
-    # Group/team mapping configuration
-    group_mapping: Mapped[dict] = mapped_column(
-        JSONB,
-        default=dict,
-        nullable=False,
-    )
-
-    # Last successful sync (for LDAP/AD)
-    last_sync_at: Mapped[datetime | None] = mapped_column(
-        DateTime(timezone=True),
-        nullable=True,
-    )
-
-    # Relationships
-    organization: Mapped["Organization | None"] = relationship(
-        "Organization",
-        back_populates="sso_providers",
-    )
-
-    user_links: Mapped[list["UserSSOLink"]] = relationship(
-        "UserSSOLink",
-        back_populates="provider",
-        cascade="all, delete-orphan",
-    )
-
-    def __repr__(self) -> str:
-        scope = f"org:{self.org_id}" if self.org_id else "global"
-        return f"<SSOProvider(type={self.provider_type}, name={self.name}, scope={scope})>"
-
-    @property
-    def is_enterprise(self) -> bool:
-        """Check if this is an enterprise SSO provider."""
-        return self.provider_type in (
-            SSOProviderType.LDAP.value,
-            SSOProviderType.ACTIVE_DIRECTORY.value,
-            SSOProviderType.OKTA.value,
-            SSOProviderType.SAML.value,
-            SSOProviderType.MICROSOFT_ENTRA.value,
-            SSOProviderType.GOOGLE_WORKSPACE.value,
-        )
-
-    def get_config_value(self, key: str, default=None):
-        """Get a configuration value safely."""
-        return self.config.get(key, default)
-
-
-class UserSSOLink(Base):
-    """
-    Link between a user and an SSO provider.
-
-    Stores the external ID from the SSO provider.
-    A user can have multiple SSO links (e.g., both Google and GitHub).
-    """
-
-    __tablename__ = "user_sso_links"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True),
-        primary_key=True,
-        default=uuid.uuid4,
-    )
-
-    user_id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True),
-        ForeignKey("users.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-
-    provider_id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True),
-        ForeignKey("sso_providers.id", ondelete="CASCADE"),
-        nullable=False,
-        index=True,
-    )
-
-    # External ID from the SSO provider
-    external_id: Mapped[str] = mapped_column(
-        String(255),
-        nullable=False,
-    )
-
-    # External email (may differ from user's primary email)
-    external_email: Mapped[str | None] = mapped_column(
-        String(255),
-        nullable=True,
-    )
-
-    # Additional data from SSO provider (renamed from 'metadata' to avoid SQLAlchemy conflict)
-    sso_metadata: Mapped[dict] = mapped_column(
-        "metadata",  # Keep original column name in database
-        JSONB,
-        default=dict,
-        nullable=False,
-    )
-
-    # Last login via this SSO provider
-    last_login_at: Mapped[datetime | None] = mapped_column(
-        DateTime(timezone=True),
-        nullable=True,
-    )
-
-    # Unique constraint: one link per provider external ID
-    __table_args__ = (UniqueConstraint("provider_id", "external_id", name="uq_provider_external_id"),)
-
-    # Relationships
-    user: Mapped["User"] = relationship(
-        "User",
-        back_populates="sso_links",
-    )
-
-    provider: Mapped["SSOProvider"] = relationship(
-        "SSOProvider",
-        back_populates="user_links",
-    )
-
-    def __repr__(self) -> str:
-        return (
-            f"<UserSSOLink(user_id={self.user_id}, " f"provider_id={self.provider_id}, external_id={self.external_id})>"
-        )
-
-    def record_login(self) -> None:
-        """Record a login via this SSO link."""
-        self.last_login_at = now_utc()
+__all__ = ["SSOProvider", "SSOProviderType", "UserSSOLink"]
